🚧 Refactor router api firo
