+++ packages/bitcore-node/src/modules/firocoin/api/contract.ts
@@ -1,0 +1,56 @@
+import { TransactionStorage } from '../../../models/transaction';
+import { ContractStorage } from '../models/contract';
+import { TokenStorage } from '../models/token';
+import { TokenBalanceStorage } from '../models/tokenBalance';
+import express = require('express');
+const router = express.Router({ mergeParams: true });
+
+router.get('/:contractAddress', async (req, res) => {
+  let { chain, network, contractAddress } = req.params;
+  try {
+    contractAddress = contractAddress.replace('0x', '');
+    const contract = await ContractStorage.getContract({ chain, network, contractAddress });
+    if (contract) {
+      contract['balance'] = 0;
+      contract['transactions'] = await TransactionStorage.collection.countDocuments({
+        chain,
+        network,
+        $or: [{ 'receipt.events.from': contractAddress }, { 'receipt.events.to': contractAddress }],
+      });
+      contract['transfers'] = await TransactionStorage.collection.countDocuments({
+        chain,
+        network,
+        $or: [{ 'receipt.events.from': contractAddress }, { 'receipt.events.to': contractAddress }],
+        'receipt.events.type': 'transfer',
+      });
+      const tokenBalances = await TokenBalanceStorage.collection
+        .find({ chain, network, address: contractAddress })
+        .sort({ _id: -1 })
+        .toArray();
+      const tokens: any[] = [];
+      for (let tokenBalnce of tokenBalances) {
+        const contractAddress = tokenBalnce.contractAddress;
+        const token = await TokenStorage.collection.findOne({ contractAddress, chain, network });
+        if (token) {
+          tokens.push({
+            contractAddress,
+            balance: tokenBalnce.balance.toString(),
+            symbol: token.symbol,
+            name: token.name,
+          });
+        }
+      }
+      contract['tokens'] = tokens;
+      res.json(contract);
+    } else {
+      res.status(404).send(`The requested contract address ${contractAddress} could not be found.`);
+    }
+  } catch (err) {
+    res.status(500).send(err);
+  }
+});
+
+module.exports = {
+  router,
+  path: '/contract',
+};
