+++ packages/bitcore-node/src/modules/firocoin/index.ts
@@ -1,8 +1,23 @@
+import express from 'express';
 import { BaseModule } from '..';
 import { FIROStateProvider } from '../../providers/chain-state/firo/firo';
 import { VerificationPeer } from '../bitcoin/VerificationPeer';
-import { FiroRoutes } from './api/firo-routes';
 import { FIROP2PWorker } from './p2p';
+
+function bootstrap(path?: string) {
+  const fs = require('fs');
+  const router = express.Router({
+    mergeParams: true,
+  });
+  const folder = path ? path + '/' : '';
+  fs.readdirSync(__dirname + '/' + path).forEach(function (file: string) {
+    if (file.match(/\.js$/) !== null && file !== 'index.js') {
+      var route = require('./' + folder + file);
+      router.use(route.path, route.router);
+    }
+  });
+  return router;
+}
 
 export default class FIROModule extends BaseModule {
   constructor(services) {
@@ -11,6 +26,6 @@ export default class FIROModule extends BaseModule {
     services.P2P.register('FIRO', FIROP2PWorker);
     services.CSP.registerService('FIRO', new FIROStateProvider());
     services.Verification.register('FIRO', VerificationPeer);
-    services.Api.app.use(FiroRoutes);
+    services.Api.app.use('/api/:chain/:network', bootstrap('api'));
   }
 }
